Remove deprecated api calls
